Refactoring that implements the subscription credit card form component and make it reusable #520
