+++ client/subscriptions/forms/index.js
@@ -1,0 +1,1 @@
+export { default as CreditCardForm } from './credit-card-form'
